moved to standardLabel namespace and added URL testing in background
began prepping for the Cloud Identity Summit Demo. I'd like to get the
system more flexible, driving the label off of data, based on the
redirect. The data will be local, but it will be a start towards the
networked system. I'm still focusing on OAuth from Facebook, but HTML
POST shouldn't be too far away. The biggest issue with all of these is
the lookup time for checking URLs and/or the DOM.  It'll suck to add a
huge delay to every single page. Ultimately, this will probably need to
be optimized in a native extension. Sigh.