flashing animation with tabids and windows
this works, but it needs severe cleaning up. first, though, I'm going to
clean it out so I don't animate it. The trick was passing the tabId to
setIcon, which handled the multiple windows automatically. But if we
clear out the animation, we can just change the icon based on whether or
not we have a standardlabel for the page. Much simpler. No intervals, no
window / tab tracking.