minimal dynamic popups with OAuth FB extraction
Simplified icon indicator (no animation, just active/default icons),
with differentiated popup windows based on whether or not we are on a FB
OAuth page, including passing the Redirect URI to the page for
processing. The Label is not currently rendered, but it has the
information it will need. We are currently passing the data as a
URIEncoded bit of JSON, so we can send just about anything to the
popup.html